SAGE-746: skip the first boot media for non-slot A boot chain
To faciliate using media redundancy for the kernel, kernel DTS, and rootfs
the primary boot media is skipped for non-slot A boot chains. This allows
the primary boot media (ex. eMMC) to be used for slot A boot chains and
for a secondary boot media (ex. SD card) to be used for slot B boot chains.